Trait pallet_dip_provider::pallet::Config

pub trait Config: Config {
    type CommitOriginCheck: EnsureOriginWithArg<Self::RuntimeOrigin, Self::Identifier, Success = Self::CommitOrigin>;
    type CommitOrigin: SubmitterInfo<Submitter = Self::AccountId>;
    type Identifier: Parameter + MaxEncodedLen;
    type IdentityCommitmentGenerator: IdentityCommitmentGenerator<Self>;
    type IdentityProvider: IdentityProvider<Self>;
    type ProviderHooks: ProviderHooks<Self>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type CommitOriginCheck: EnsureOriginWithArg<Self::RuntimeOrigin, Self::Identifier, Success = Self::CommitOrigin>

The check ensuring a given runtime origin is allowed to generate and remove identity commitments.

type CommitOrigin: SubmitterInfo<Submitter = Self::AccountId>

The resulting origin if CommitOriginCheck returns with errors. The origin is not required to be an AccountId, but must include information about the AccountId of the tx submitter.

type Identifier: Parameter + MaxEncodedLen

The type of an identifier used to retrieve identity information about a subject.

type IdentityCommitmentGenerator: IdentityCommitmentGenerator<Self>

The type responsible for generating identity commitments, given the identity information associated to a given Identifier.

type IdentityProvider: IdentityProvider<Self>

The type responsible for retrieving the information associated to a subject given their identifier. The information can potentially be retrieved from any source, using a combination of on-chain and off-chain solutions.

type ProviderHooks: ProviderHooks<Self>

Customizable external logic to handle events in which a new identity commitment is generated or removed.

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The aggregate Event type.

type WeightInfo: WeightInfo